requirements.txt
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.50.0
pandas>=1.1.0
scikit-learn>=0.23.0
nlp>=0.4.0